+++ src/modules/blockchainApplication/connection/components/RequestView/index.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import MultiStep from 'src/modules/common/components/OldMultiStep';
+import MultiStep from '@common/components/OldMultiStep';
 import TxSignatureCollector from '@transaction/components/TxSignatureCollector';
 import Dialog from 'src/theme/dialog/dialog';
-import Summary from '@wallet/components/SignWCSummary';
-import Status from '@wallet/components/SignWCStatus';
+import Summary from '@wallet/components/RequestSignSummary';
+import Status from '@wallet/components/RequestSignStatus';
 import RequestSummary from '../RequestSummary';
 import RequestView from './RequestView';
