+++ src/modules/wallet/components/RequestSignStatus/index.test.js
@@ -3,7 +3,7 @@ import { act } from 'react-dom/test-utils';
 import * as reactRedux from 'react-redux';
 import { useSession } from '@libs/wcm/hooks/useSession';
 import { renderWithRouterAndQueryClient } from 'src/utils/testHelpers';
-import SignWCStatus from './index';
+import RequestSignStatus from './index';
 
 jest.mock('@walletconnect/utils', () => ({
   getSdkError: jest.fn((str) => str),
@@ -46,7 +46,7 @@ const failureTransactions = {
   txSignatureError: 'error',
 };
 
-describe('SignWCStatus', () => {
+describe('RequestSignStatus', () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
@@ -54,7 +54,7 @@ describe('SignWCStatus', () => {
   it('render the signature success result', () => {
     reactRedux.useSelector.mockReturnValue(successTransactions);
     useSession.mockReturnValue({ respond, reject });
-    renderWithRouterAndQueryClient(SignWCStatus, props);
+    renderWithRouterAndQueryClient(RequestSignStatus, props);
     expect(screen.getByText('Transaction signing successful')).toBeInTheDocument();
     expect(screen.getByText('Your transaction has been signed, click the button below to copy your signed transaction, once copied you will be redirected to application.')).toBeInTheDocument();
     expect(screen.getByText('Copy and return to application')).toBeInTheDocument();
@@ -64,7 +64,7 @@ describe('SignWCStatus', () => {
   it('render the signature failure result', () => {
     reactRedux.useSelector.mockReturnValue(failureTransactions);
     useSession.mockReturnValue({ respond, reject });
-    renderWithRouterAndQueryClient(SignWCStatus, props);
+    renderWithRouterAndQueryClient(RequestSignStatus, props);
     expect(screen.getByText('Transaction signing failed')).toBeInTheDocument();
     expect(screen.getByText('There was an error signing your transaction. please close this dialog and try again.')).toBeInTheDocument();
     expect(screen.queryAllByText('Copy and return to application')).toHaveLength(0);
@@ -74,7 +74,7 @@ describe('SignWCStatus', () => {
   it('copy the signature if clicked on the copy button', () => {
     reactRedux.useSelector.mockReturnValue(successTransactions);
     useSession.mockReturnValue({ respond });
-    renderWithRouterAndQueryClient(SignWCStatus, props);
+    renderWithRouterAndQueryClient(RequestSignStatus, props);
     expect(screen.getByText('Copy and return to application')).toBeInTheDocument();
     const button = screen.getAllByRole('button')[0];
     act(() => {
@@ -90,7 +90,7 @@ describe('SignWCStatus', () => {
   it('reject the signature', () => {
     reactRedux.useSelector.mockReturnValue(successTransactions);
     useSession.mockReturnValue({ reject });
-    renderWithRouterAndQueryClient(SignWCStatus, props);
+    renderWithRouterAndQueryClient(RequestSignStatus, props);
     const button = screen.getAllByRole('button')[1];
     act(() => {
       fireEvent.click(button);
